passport-oauth2 implementation for orcid
This commit is pretty far along the implementation for orcid
authentication. Sadly we hit a roadblock where the orcid_id in the
request for a access token cannot be retrieved through the passport
functions and is lost. we haven't found a way to later on retrieve this
orcid_id or the profile with the accesstoken through the api.

If possible, this should be done in the passport.userProfile() function,
which would then pass this profile on in the whole workflow, saving it
to the mongodb.

implementation:
/auth/orcid will start the oauth2 workflow and authenticate against the
orcid public api. as soon as the user is authenticated via oauth2, a
ookie is saved through express-session which will contain a reference to
the session & user details in the mongodb collection 'Sessions'. this is
the place where the orcid_id/userid will also be saved to and can be
accessed from other services.